Huge type rename: Short -> Int16, Int -> Int32, Long -> Int64, Float -> Float32, Double -> Float64

+++ spec/ast/to_s_spec.rb
@@ -2,11 +2,11 @@ require 'spec_helper'
 
 describe "ast nodes" do
   it "should to_s Int" do
-    5.int.to_s.should eq('5')
+    5.int32.to_s.should eq('5')
   end
 
   it "should to_s Float" do
-    5.0.float.to_s.should eq('5.0')
+    5.0.float64.to_s.should eq('5.0')
   end
 
   it "should to_s True" do
@@ -29,20 +29,20 @@ describe "ast nodes" do
     ">=",
   ].each do |op|
     it "should to_s Call #{op}" do
-      Call.new(5.int, op.to_sym, [6.int]).to_s.should eq("5 #{op} 6")
+      Call.new(5.int32, op.to_sym, [6.int32]).to_s.should eq("5 #{op} 6")
     end
   end
 
   it "should to_s Def with no args" do
-    Def.new("foo", [], [1.int]).to_s.should eq("def foo\n  1\nend")
+    Def.new("foo", [], [1.int32]).to_s.should eq("def foo\n  1\nend")
   end
 
   it "should to_s Def with args" do
-    Def.new("foo", ['var'.arg], [1.int]).to_s.should eq("def foo(var)\n  1\nend")
+    Def.new("foo", ['var'.arg], [1.int32]).to_s.should eq("def foo(var)\n  1\nend")
   end
 
   it "should to_s Def with many expressions" do
-    Def.new("foo", [], [1.int, 2.int]).to_s.should eq("def foo\n  1\n  2\nend")
+    Def.new("foo", [], [1.int32, 2.int32]).to_s.should eq("def foo\n  1\n  2\nend")
   end
 
   it "should to_s Var" do
@@ -54,7 +54,7 @@ describe "ast nodes" do
   end
 
   it "should to_s Call with args" do
-    Call.new(nil, "foo", [1.int, 2.int]).to_s.should eq("foo(1, 2)")
+    Call.new(nil, "foo", [1.int32, 2.int32]).to_s.should eq("foo(1, 2)")
   end
 
   it "should to_s Call with no block" do
@@ -62,7 +62,7 @@ describe "ast nodes" do
   end
 
   it "should to_s If" do
-    If.new("foo".var, 1.int).to_s.should eq("if foo\n  1\nend")
+    If.new("foo".var, 1.int32).to_s.should eq("if foo\n  1\nend")
   end
 
   it "should to_s Not" do
@@ -71,7 +71,7 @@ describe "ast nodes" do
 
   ['return', 'break', 'next', 'yield'].each do |keyword|
     it "should to_s #{keyword.capitalize}" do
-      eval(keyword.capitalize).new(["foo".var, 1.int]).to_s.should eq("#{keyword} foo, 1")
+      eval(keyword.capitalize).new(["foo".var, 1.int32]).to_s.should eq("#{keyword} foo, 1")
     end
   end
 end
